% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_wrappers.R
\name{annotate_bmkc_drivers}
\alias{annotate_bmkc_drivers}
\title{Annotate Master Regulators for BMKC Modules}
\usage{
annotate_bmkc_drivers(bmkc_result, tf_list)
}
\arguments{
\item{bmkc_result}{The output object from `run_bmkc()`.}

\item{tf_list}{A character vector of known Transcription Factor gene symbols.}
}
\value{
An updated `bmkc_result` object with a new `$module_drivers` list.
}
\description{
Scans BMKC co-expression modules to identify the embedded Transcription Factors acting as clique master regulators.
}
