% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_wrappers.R
\name{annotate_rosdet_direction}
\alias{annotate_rosdet_direction}
\title{Annotate Direction for ROS-DET Switching Pairs}
\usage{
annotate_rosdet_direction(rosdet_result, tf_list)
}
\arguments{
\item{rosdet_result}{The output object from `run_rosdet()`.}

\item{tf_list}{A character vector of known Transcription Factor gene symbols.}
}
\value{
An updated `rosdet_result` object with a new `$Direction` column.
}
\description{
Assigns directed edges to ROS-DET switching pairs based on prior Transcription Factor knowledge.
}
