% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphics.R
\name{bicor_palette}
\alias{bicor_palette}
\title{Standard Palette for bicorDCEA}
\usage{
bicor_palette()
}
\description{
Standardized hex codes to keep Condition 1 and Condition 2 visually consistent across all plots.
}
