% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{biweight_midvariance}
\alias{biweight_midvariance}
\title{Calculate Biweight Midvariance}
\usage{
biweight_midvariance(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The biweight midvariance of the vector.
}
\description{
A robust measure of variance based on the median absolute deviation.
This is a helper function for ROS-DET's weighting step.
}
\keyword{internal}
