% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_cor_matrix}
\alias{compute_cor_matrix}
\title{Compute full bicor correlation matrix (internal)}
\usage{
compute_cor_matrix(mat, workers = 1)
}
\arguments{
\item{mat}{Numeric matrix (genes × samples)}

\item{workers}{Integer. Number of parallel workers}
}
\value{
Symmetric numeric matrix with gene names
}
\description{
Compute full bicor correlation matrix (internal)
}
\keyword{internal}
