% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_wrappers.R
\name{export_cytoscape}
\alias{export_cytoscape}
\title{Export ROS-DET Network for Cytoscape}
\usage{
export_cytoscape(rosdet_result, output_dir = "Cytoscape_Networks/")
}
\arguments{
\item{rosdet_result}{A `rosdet_result` object from [run_rosdet()].}

\item{output_dir}{Directory to write the node and edge tables to.
Created if it does not exist. Default `"Cytoscape_Networks/"`.}
}
\value{
Invisibly, the output directory path. Called for its side
  effect of writing Cytoscape-compatible node and edge tables.
}
\description{
Export ROS-DET Network for Cytoscape
}
