% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_bmht_plots.R
\name{plot_3d_network}
\alias{plot_3d_network}
\title{Plot 3D Network Landscape}
\usage{
plot_3d_network(bmht_result, n_top = 20, edge_threshold = 0.5)
}
\description{
Renders an interactive 3D topography using plotly.
}
