% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmkc.R
\name{plot_bmkc_exploration}
\alias{plot_bmkc_exploration}
\title{Plot the results of a BMKC threshold exploration.}
\usage{
plot_bmkc_exploration(exploration_result)
}
\arguments{
\item{exploration_result}{A data frame generated by `explore_bmkc_thresholds`.}
}
\value{
A ggplot2 object.
}
\description{
Generates a heatmap visualizing the number of modules found for different
combinations of T1 and T2 thresholds.
}
