% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphics.R
\name{plot_enrichment_dotplot}
\alias{plot_enrichment_dotplot}
\title{GO Enrichment Dot Plot wrapper}
\usage{
plot_enrichment_dotplot(enrich_res, top_n = 15, title = "Pathway Enrichment")
}
\description{
A clean, standardized dot plot for displaying pathway enrichment results of network hubs.
}
