% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_bmht_plots.R
\name{plot_hive_network}
\alias{plot_hive_network}
\title{Plot Network Hive Layout (BMHT Hubs)}
\usage{
plot_hive_network(bmht_result, n_top = 30, min_edge = 0.4)
}
\description{
A highly structured radial layout for displaying massive hubs without edge overlapping.
}
