% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_rosdet_plots.R
\name{plot_intercellular_network}
\alias{plot_intercellular_network}
\title{Plot Circular intercellular Diagram (ROS-DET)}
\usage{
plot_intercellular_network(rosdet_result, n_top = 30)
}
\arguments{
\item{rosdet_result}{A ROS-DET result object.}

\item{n_top}{Numeric. The maximum number of top switching edges to plot. Default is 30.}
}
\description{
Circular layout perfect for Single-Cell Communication and massive Host-Pathogen interactions.
}
