% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphics.R
\name{plot_method_overlap}
\alias{plot_method_overlap}
\title{Multi-Method UpSet Plot (Intersection of Discoveries)}
\usage{
plot_method_overlap(rosdet_res = NULL, bmht_res = NULL, bmkc_res = NULL)
}
\description{
Compares the genes identified by ROS-DET, BMHT, and BMKC to find universal network drivers.
}
