% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_rosdet_plots.R
\name{plot_rosdet_quadrant}
\alias{plot_rosdet_quadrant}
\title{The Quadrant Scatter (ROS-DET "Flip-Flop" Proof)}
\usage{
plot_rosdet_quadrant(rosdet_result, n_label = 15)
}
\description{
Plots Cond1 vs Cond2 correlation to visualize the complete inversion.
}
