% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_rosdet_plots.R
\name{plot_species_interaction}
\alias{plot_species_interaction}
\title{Species Bipartite Interaction Network (ROS-DET)}
\usage{
plot_species_interaction(rosdet_result)
}
\description{
Creates a bipartite network for Host-Pathogen / Cross-Species modes.
}
