% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_rosdet_plots.R
\name{plot_switching_stacked}
\alias{plot_switching_stacked}
\title{Stacked Switching Diagram (ROS-DET "Brick" Plot)}
\usage{
plot_switching_stacked(rosdet_result, n_top = 12)
}
\description{
Shows how many specific relationships completely inverted their logic.
}
