% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_bmht_plots.R
\name{plot_top_gene_bar}
\alias{plot_top_gene_bar}
\title{Top Gene Connectivity Bar Chart (BMHT)}
\usage{
plot_top_gene_bar(bmht_result, n_top = 15)
}
\description{
Renders a bar chart of the most heavily rewired genes.
}
