% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_bmkc_plots.R
\name{plot_umap_eigengene}
\alias{plot_umap_eigengene}
\title{Module Eigengene Projected onto a UMAP Embedding}
\usage{
plot_umap_eigengene(
  bmkc_res,
  sc_obj,
  module_index = 1,
  reduction = "umap",
  pt_size = 0.5,
  split_by = NULL,
  ...
)
}
\arguments{
\item{bmkc_res}{A `bmkc_result` object from [run_bmkc()].}

\item{sc_obj}{A normalized Seurat object containing a UMAP (or other)
dimensionality reduction.}

\item{module_index}{Which detected module to score. Default `1`.}

\item{reduction}{Name of the dimensionality reduction to plot.
Default `"umap"`.}

\item{pt_size}{Point size. Default `0.5`.}

\item{split_by}{Optional metadata column to facet the plot by.}

\item{...}{Currently unused; accepted for S3 compatibility.}
}
\value{
A `ggplot` object.
}
\description{
Module Eigengene Projected onto a UMAP Embedding
}
