% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmht.R
\name{summary.bmht_result}
\alias{summary.bmht_result}
\title{Print a summary of BMHT results.}
\usage{
\method{summary}{bmht_result}(object, fdr_threshold = 0.05, ...)
}
\arguments{
\item{object}{An object of class `bmht_result`.}

\item{fdr_threshold}{The FDR cutoff to use for counting significant genes. Defaults to 0.05.}

\item{...}{Additional arguments (not used).}
}
\value{
Prints a summary to the console, invisibly returning the object.
}
\description{
Provides a concise summary of the results from a `run_bmht` analysis.
}
