% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosdet.R
\name{summary.rosdet_result}
\alias{summary.rosdet_result}
\title{Print a summary of ROS-DET results.}
\usage{
\method{summary}{rosdet_result}(object, ...)
}
\arguments{
\item{object}{An object of class `rosdet_result`.}

\item{...}{Additional arguments (not used).}
}
\value{
Prints a summary to the console, invisibly returning the object.
}
\description{
Provides a concise summary of the results from a `run_rosdet` analysis.
}
