% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_wrappers.R
\name{annotate_rosdet_direction}
\alias{annotate_rosdet_direction}
\title{Annotate Direction for ROS-DET Switching Pairs}
\usage{
annotate_rosdet_direction(rosdet_result, tf_list)
}
\arguments{
\item{rosdet_result}{A `rosdet_result` object from [run_rosdet()].}

\item{tf_list}{Character vector of transcription factor (or other
regulator) gene names.}
}
\value{
The results data frame with added `G1_is_TF`, `G2_is_TF` and
  `Direction` columns indicating putative regulatory direction.
}
\description{
Annotate Direction for ROS-DET Switching Pairs
}
