% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bicor}
\alias{bicor}
\title{Calculate Biweight Midcorrelation (bicor)}
\usage{
bicor(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
The biweight midcorrelation coefficient.
}
\description{
A robust measure of correlation that is less sensitive to outliers than
Pearson correlation. This is an independent implementation based on the
formulas described by Kayano et al. (2011), Zheng et al. (2014),
and Yuan et al. (2015).
}
\details{
Calculate Biweight Midcorrelation (bicor)

A robust measure of correlation that is less sensitive to outliers.
}
