% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_wrappers.R
\name{export_cytoscape}
\alias{export_cytoscape}
\title{Export ROS-DET Network for Cytoscape}
\usage{
export_cytoscape(rosdet_result, output_dir = "Cytoscape_Networks/")
}
\arguments{
\item{rosdet_result}{The output object from `run_rosdet()`, preferably annotated with directions.}

\item{output_dir}{String. The directory to save the CSV files.}
}
\description{
Generates Cytoscape-ready Node and Edge tables from a ROS-DET result object.
}
