% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmht.R
\name{plot.bmht_result}
\alias{plot.bmht_result}
\title{Plot the top-ranked genes from a BMHT analysis.}
\usage{
\method{plot}{bmht_result}(x, n_top = 15, ...)
}
\arguments{
\item{x}{An object of class `bmht_result`.}

\item{n_top}{The number of top genes to display in the plot. Defaults to 15.}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot2 object.
}
\description{
Generates a bar plot showing the top differentially co-expressed genes,
ranked by their DC_Score.
}
