% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_bmkc_plots.R
\name{plot.bmkc_result}
\alias{plot.bmkc_result}
\title{Plot Method for BMKC}
\usage{
\method{plot}{bmkc_result}(
  x,
  type = c("default", "network", "heatmap", "eigengene", "raincloud", "spatial",
    "density", "overlap", "umap"),
  ...
)
}
\arguments{
\item{x}{A bmkc_result object.}

\item{type}{String. Type of plot: "network", "heatmap", "eigengene", "raincloud", "spatial", "density", "overlap", or "umap".}

\item{...}{Additional arguments passed to the specific plotting function.}
}
\description{
S3 method to easily route plotting commands based on type.
}
