% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_rosdet_plots.R
\name{plot.rosdet_result}
\alias{plot.rosdet_result}
\title{Plot Method for ROS-DET}
\usage{
\method{plot}{rosdet_result}(
  x,
  type = c("default", "quadrant", "stacked", "species", "volcano", "chord"),
  ...
)
}
\arguments{
\item{x}{A rosdet_result object.}

\item{type}{String. Type of plot: "default", "quadrant", "stacked", "species", "volcano", or "chord".}

\item{...}{Additional arguments passed to the specific plotting function.}
}
\description{
S3 method to easily route plotting commands based on type.
}
