% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosdet.R
\name{plot.rosdet_result}
\alias{plot.rosdet_result}
\title{Plot the top switching gene pair from a ROS-DET analysis.}
\usage{
\method{plot}{rosdet_result}(x, ...)
}
\arguments{
\item{x}{An object of class `rosdet_result`.}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot2 object.
}
\description{
Generates a scatter plot visualizing the correlation switch for the
top-ranked gene pair identified by `run_rosdet`.
}
