% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_rosdet_plots.R
\name{plot_edge_volcano}
\alias{plot_edge_volcano}
\title{Plot Differential Edge Volcano}
\usage{
plot_edge_volcano(rosdet_res)
}
\description{
Plots the change in correlation (Delta r) vs significance for gene pairs.
}
