% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_bmkc_plots.R
\name{plot_module_eigengene}
\alias{plot_module_eigengene}
\title{Module Eigengene Boxplot (BMKC)}
\usage{
plot_module_eigengene(bmkc_result, module_index = 1)
}
\description{
Plots the summary expression (First Principal Component) of the entire module.
}
