% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_bmht_plots.R
\name{plot_top_gene_bar}
\alias{plot_top_gene_bar}
\title{Top Gene Connectivity Bar Chart (BMHT)}
\usage{
plot_top_gene_bar(bmht_result, n_top = 15, ...)
}
\arguments{
\item{bmht_result}{A `bmht_result` object from [run_bmht()].}

\item{n_top}{Number of top-ranked genes to display. Default `15`.}

\item{...}{Currently unused; accepted for S3 compatibility.}
}
\value{
A `ggplot` object.
}
\description{
Top Gene Connectivity Bar Chart (BMHT)
}
