% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphics.R
\name{plot_trajectory_heatmap}
\alias{plot_trajectory_heatmap}
\title{Plot Rolling Trajectory Heatmap}
\usage{
plot_trajectory_heatmap(
  expr_matrix,
  geneA,
  geneB,
  pseudotime_vec,
  window_size = 30
)
}
\description{
For scRNA-seq trajectory analysis. Plots how the correlation between two genes changes dynamically over pseudotime.
}
