% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphics.R
\name{theme_bicor}
\alias{theme_bicor}
\title{Custom bicorX ggplot2 Theme}
\usage{
theme_bicor(base_size = 14)
}
\arguments{
\item{base_size}{Base font size in points. Default `14`.}
}
\value{
A `ggplot2` theme object.
}
\description{
Custom bicorX ggplot2 Theme
}
